collect codes about notices
